@@ -1,10 +1,10 @@
-ruby '3.0.3'
+ruby '3.2.4'
 source 'https://rubygems.org'
 gem 'rack', '~> 2.2'
 gem 'rspec', '~> 3.11'
 gem 'rack-test', '~> 2.0'
 gem 'sinatra', '~> 2.2'
 gem 'dotenv', '~> 2.8'
-gem 'pg', '~> 1.4'
+gem 'pg', '~> 1.5'
 gem 'webrick', '~> 1.7'
 gem "activerecord", "~> 7.0"
