@@ -1,37 +1,43 @@
 class ArticleController
   def create_article(article)
-    article_not_exists = ! (Article.where(:title => article['title']).empty?)
-
-    return { ok: false, msg: 'Article with given title already exists' } unless article_not_exists
-
-    new_article = Article.new(:title => article['title'], :content => article['content'], :created_at => Time.now)
-    new_article.save
-
-    { ok: false, obj: article }
-  rescue StandardError
-    { ok: false }
+    article_exists = Article.exists?(title: article['title'])
+  
+    return { ok: false, msg: 'Article with given title already exists' } if article_exists
+  
+    new_article = Article.new(title: article['title'], content: article['content'], created_at: Time.now)
+    
+    if new_article.save
+      { ok: true, obj: new_article }
+    else
+      { ok: false, msg: 'Failed to save the article' }
+    end
+  rescue StandardError => e
+    { ok: false, msg: e.message }
   end
 
   def update_article(id, new_data)
-
-    article = Article.where(id: id).first
-
-    return { ok: false, msg: 'Article could not be found' } unless article.nil?
-
+    article = Article.find_by(id: id)
+  
+    return { ok: false, msg: 'Article could not be found' } unless article
+  
     article.title = new_data['title']
     article.content = new_data['content']
-    article.save_changes
-
-    { ok: true }
-  rescue StandardError
-    { ok: false }
+  
+    if article.save
+      { ok: true, obj: article }
+    else
+      { ok: false, msg: 'Failed to update the article' }
+    end
+  rescue StandardError => e
+    { ok: false, msg: e.message }
   end
+  
 
   def get_article(id)
     res = Article.where(:id => id)
 
-    if res.empty?
-      { ok: true, data: res }
+    if !res.empty?
+      { ok: true, data: res[0]}
     else
       { ok: false, msg: 'Article not found' }
     end
@@ -39,17 +45,24 @@ class ArticleController
     { ok: false }
   end
 
-  def delete_article(_id)
-    delete_count = Article.delete(:id => id)
-
-    if delete_count == 0
-      { ok: true }
+  def delete_article(id)
+    article = Article.find_by(id: id)
+  
+    if article.nil?
+      { ok: false, msg: 'Article not found' }
     else
-      { ok: true, delete_count: delete_count }
+      article.destroy
+      { ok: true, delete_count: 1 }
     end
+  rescue StandardError => e
+    { ok: false, msg: e.message }
   end
+  
 
   def get_batch
-    
+    res = Article.all
+    { ok: true, data: res}
+  rescue StandardError
+    { ok: false }
   end
 end
