@@ -1,7 +1,7 @@
 class HealthRoutes < Sinatra::Base
     
   get('/') do
-    if request.env['AUTHED'] == true
+    if request.env['AUTHED'] = true
       'App working OK'
     else
       status 403
